Split to_hats tests to separate file.
